# api/openapi.yaml
openapi: 3.0.3
info:
  title: service-registry
  version: 1.0.0
  description: |
    全局独立的**服务注册中心**：服务契约（含对外 API 这一基础属性）与实例元信息的
    唯一存储中心。其他平台**从本中心拉取**服务元信息（快照 / 增量游标 / SSE）。

    **语义边界：本中心只记录"登记了什么"——不探活、不心跳、不启停任何进程，
    也不保证实例可达。** 可达性由消费方自行校验（契约里的 `healthPath` 只是元信息）。

    鉴权：
      - 写接口（注册/维护/管理）需要令牌：`Authorization: Bearer <token>` 或 `X-Registry-Token`。
        `admin token` 全权；`namespace token` 只能写自己那个命名空间。
      - 读接口默认开放（便于各平台拉取），可用 `REGISTRY_READ_AUTH=token` 收紧。
      - 未配置 admin token 时是**本机开发模式**：写接口不鉴权。
    error 响应统一为 `{"error":{"code":"...","message":"..."}}`。
servers:
  - url: http://127.0.0.1:4240
...
paths:
...
  /v1/events:
    get:
      tags: [pull]
      summary: 实时变更订阅（SSE）
      description: |
        `text/event-stream` 流：
          - `event: hello`  建立连接时的当前 revision
          - `event: change` 每一条变更（与 /v1/changes 同构）
          - `: ping`        保活注释
        参数 `since` 可指定起始游标；`namespace` / `service` 可过滤。
      parameters:
        - { in: query, name: since, schema: { type: integer, format: int64 } }
        - { in: query, name: namespace, schema: { type: string } }
        - { in: query, name: service, schema: { type: string } }
      responses:
        '200':
          description: SSE 流
          content:
            text/event-stream: { schema: { type: string } }
...
  /v1/services:
    get:
      tags: [discover]
      summary: 服务目录（可跨命名空间检索）
      parameters:
        - { in: query, name: namespace, schema: { type: string } }
        - { in: query, name: tag, schema: { type: string } }
        - { in: query, name: owner, schema: { type: string } }
        - { in: query, name: protocol, schema: { type: string } }
        - { in: query, name: q, schema: { type: string }, description: 关键字（名字/描述/owner/version） }
        - { in: query, name: limit, schema: { type: integer, default: 200, maximum: 2000 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: 服务列表（含端点索引与实例数）
          content:
            application/json:
              schema:
                type: object
                properties:
                  services: { type: array, items: { $ref: '#/components/schemas/Service' } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
  /v1/namespaces/{ns}/services:
    get:
      tags: [discover]
      summary: 某命名空间下的服务目录
      parameters:
        - $ref: '#/components/parameters/ns'
        - { in: query, name: tag, schema: { type: string } }
        - { in: query, name: q, schema: { type: string } }
      responses:
        '200': { description: 同 /v1/services }
...
  /v1/search/apis:
    get:
      tags: [discover]
      summary: 反向查找——"这个接口谁提供"
      description: |
        `path` 支持三种写法：
          1. 具体路径 `/v1/streams/abc/events` → 命中登记的模板 `/v1/streams/{stream}/events`（matchType=template）
          2. 登记模板原样 `/v1/streams/{stream}/events` → matchType=exact
          3. 通配 `/v1/streams/**` → `*` 匹配单段、`**` 跨段（matchType=glob）

        默认按 精确 → 模板 → 通配 依次尝试；`match=` 可强制指定其中一种。
      parameters:
        - { in: query, name: method, schema: { type: string, example: GET } }
        - { in: query, name: path, schema: { type: string } }
        - { in: query, name: match, schema: { type: string, enum: [exact, template, glob] } }
        - { in: query, name: namespace, schema: { type: string } }
        - { in: query, name: tag, schema: { type: string }, description: 端点标签 }
        - { in: query, name: q, schema: { type: string }, description: 关键字（服务名/描述/摘要/路径） }
        - { in: query, name: limit, schema: { type: integer, default: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: 命中列表（精确/模板命中排在通配之前）
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches: { type: array, items: { $ref: '#/components/schemas/EndpointMatch' } }
                  total: { type: integer }
                  hasMore: { type: boolean }
                  query: { type: object }
...
  /v1/namespaces/{ns}/services/{svc}:
    parameters:
      - $ref: '#/components/parameters/ns'
      - $ref: '#/components/parameters/svc'
    get:
      tags: [discover]
      summary: 服务契约详情（含对外 API 端点索引与实例数）
      responses:
        '200':
          description: 契约
          content:
            application/json:
              schema: { type: object, properties: { service: { $ref: '#/components/schemas/Service' } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [register]
      summary: 登记或更新服务契约（幂等）
      description: |
        **服务的对外 API 是本注册中心的基础属性**：请求体必须提供
        `api.spec`（内联 OpenAPI 3.x / Swagger 2.0 原文，YAML 或 JSON）或
        `api.endpoints`（显式端点列表），否则 400。

        提供 `api.spec` 时：服务端解析出端点索引、计算 `specHash` 并原样保存原文
        （可通过 `.../spec` 取回）；`api.endpoints` 优先于从 spec 解析的结果。

        端点索引是**整体替换**：本次没写进去的端点不会残留。
        实例地址不在这个接口登记，走 `.../instances`。
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ServiceWrite' } }
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  service: { $ref: '#/components/schemas/Service' }
                  created: { type: boolean }
        '200':
          description: 已更新（幂等重复登记）
          content:
            application/json:
              schema:
                type: object
                properties:
                  service: { $ref: '#/components/schemas/Service' }
                  created: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      tags: [register]
      summary: 删除服务契约（连带其端点索引与全部实例）
      responses:
        '204': { description: 已删除 }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    ns:
      in: path
      name: ns
      required: true
      schema: { type: string, pattern: '^[a-z0-9]([a-z0-9._-]{0,61}[a-z0-9])?$' }
      description: 命名空间
    svc:
      in: path
      name: svc
      required: true
      schema: { type: string, pattern: '^[a-z0-9]([a-z0-9._-]{0,61}[a-z0-9])?$' }
      description: 服务名
  responses:
    BadRequest:
      description: 请求不合法（字段校验失败、API 来源缺失、规格无法解析等）
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    NotFound:
      description: 对象不存在
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: not_found }
            message: { type: string }
    AuthScheme:
      type: object
      description: 服务对外 API 的鉴权方式（元信息，仅用于展示与检索）
      properties:
        scheme: { type: string, example: bearer }
        in: { type: string, enum: [header, query, cookie] }
        name: { type: string, example: Authorization }
        notes: { type: string }
    Endpoint:
      type: object
      description: 服务对外 API 的一个端点
      properties:
        method: { type: string, example: GET }
        path: { type: string, example: '/v1/streams/{stream}/events' }
        summary: { type: string }
        operationId: { type: string }
        tags: { type: array, items: { type: string } }
        auth: { type: array, items: { type: string }, description: 引用的 securityScheme 名 }
    ServiceAPI:
      type: object
      description: 服务的对外 API 描述（spec 原文不在这里返回，走 .../spec）
      properties:
        protocols: { type: array, items: { type: string }, example: [http] }
        authSchemes: { type: array, items: { $ref: '#/components/schemas/AuthScheme' } }
        docsUrl: { type: string }
        specUrl: { type: string, description: 外部规范链接（只存不抓取） }
        specHash: { type: string, description: 内联规范原文的 sha256 }
        specBytes: { type: integer }
        hasSpec: { type: boolean }
        endpoints: { type: array, items: { $ref: '#/components/schemas/Endpoint' } }
    Service:
      type: object
      description: 服务契约（低频变更）
      properties:
        namespace: { type: string }
        name: { type: string }
        version: { type: string }
        owner: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        basePath: { type: string }
        healthPath: { type: string, description: 元信息：消费方/看门狗可据此自行探活 }
        api: { $ref: '#/components/schemas/ServiceAPI' }
        revision: { type: integer, format: int64, description: 最后一次被改时的全局 revision }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        registeredBy: { type: string, description: 谁登记/改的（审计） }
        instanceCount: { type: integer }
    ServiceWrite:
      type: object
      description: 登记契约的请求体（与 Service 同构，可"GET → 改 → PUT"闭环）
      properties:
        version: { type: string }
        owner: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        basePath: { type: string }
        healthPath: { type: string, description: 必须以 / 开头 }
        api:
          allOf:
            - $ref: '#/components/schemas/ServiceAPI'
            - type: object
              properties:
                spec: { type: string, description: 内联 OpenAPI 3.x / Swagger 2.0 原文（YAML 或 JSON，仅写入用） }
                specFormat: { type: string, enum: [yaml, json], description: 省略时按内容自动判定 }
...
    EndpointMatch:
      type: object
      description: 一次"这个接口谁提供"的命中
      properties:
        namespace: { type: string }
        service: { type: string }
        serviceVersion: { type: string }
        description: { type: string }
        instanceCount: { type: integer }
        matchType: { type: string, enum: [exact, template, glob] }
        endpoint: { $ref: '#/components/schemas/Endpoint' }
...